requirements.txt
alembic==1.16.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
dotenv==0.9.9
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysqlclient==2.2.7
numpy==2.2.5
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
platformdirs==4.3.7
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
virtualenv==20.30.0
waitress==3.0.2
Werkzeug==3.1.3
wfastcgi==3.0.0
